src/test/java/net/mloeks/aoc22/FoodServiceTest.java
package net.mloeks.aoc22;

import org.junit.jupiter.api.Test;

import java.io.IOException;
import java.net.URISyntaxException;
import java.util.List;

import static java.util.Comparator.comparingInt;
import static org.assertj.core.api.Assertions.assertThat;

public class FoodServiceTest {

    private final FoodService cut = new FoodService();

    @Test
    public void day1() throws IOException, URISyntaxException {
        List<Integer> caloriesCarriedByElf = cut.getCaloriesCarriedByElves("01_food_calories.txt");
        caloriesCarriedByElf.sort(comparingInt(value -> (int) value).reversed());

        // part 1
        int maxCaloriesCarriedByElf = caloriesCarriedByElf.get(0);
        assertThat(maxCaloriesCarriedByElf).isEqualTo(71924);

        // part 2
        int topThreeElvesCaloriesSum = caloriesCarriedByElf.get(0) + caloriesCarriedByElf.get(1) + caloriesCarriedByElf.get(2);
        assertThat(topThreeElvesCaloriesSum).isEqualTo(210406);
    }

}
